requirements.txt
toga-winforms==0.4.7
travertino==0.3.0
aiohttp==3.12.13
aiohttp-socks==0.10.1
py7zr==0.22.0
qrcode==8.2
pillow==11.2.1
pandas==2.3.0
flask[async]==3.1.1